+++ cloudsplaining/scan/user_details.py
@@ -205,6 +205,14 @@ class UserDetail:
         return policies
 
     @property
+    def attached_managed_policies_pointer_json(self):
+        """Return JSON representation of attached managed policies - but just with pointers to the Policy ID"""
+        policies = {}
+        for policy in self.attached_managed_policies:
+            policies[policy.policy_id] = policy.policy_name
+        return policies
+
+    @property
     def inline_policies_json(self):
         """Return JSON representation of attached inline policies"""
         policies = {}
@@ -235,7 +243,7 @@ class UserDetail:
             groups=self.groups_json,
             path=self.path,
             managed_policies_count=len(self.attached_managed_policies),
-            managed_policies=self.attached_managed_policies_json,
+            managed_policies=self.attached_managed_policies_pointer_json,
             risks=self.consolidated_risks
         )
         return this_user_detail